remove generators directory, move everything to lib
