--- generators/pattern-library.js
@@ -1,1 +1,0 @@
-module.exports = require( './pattern-library/index' );
