+++ lib/js/configure.js
@@ -3,7 +3,7 @@
 var path = require( 'path' );
 
 var _ = require( 'lodash' );
-var log = require( '../../lib/logger' );
+var log = require( './logger' );
 var util = require( 'util' );
 
 /**
