+++ lib/js/fs-promiser.js
