+++ lib/js/handlebars-helpers.js
@@ -1,4 +1,4 @@
-var log = require( '../lib/logger' );
+var log = require( './logger' );
 
 module.exports = {
