+++ lib/js/index.js
@@ -5,9 +5,9 @@ var path = require( 'path' );
 var parser = require( './parser' );
 var render = require( './render' );
 var configure = require( './configure' );
-var globber = require( '../../lib/globber' );
-var log = require( '../../lib/logger' );
-var fsp = require( '../../lib/fs-promiser' );
+var globber = require( './globber' );
+var log = require( './logger' );
+var fsp = require( './fs-promiser' );
 
 /**
  *
