+++ lib/js/logger.js
