+++ lib/js/parser.js
@@ -7,7 +7,7 @@
 
 'use strict';
 
-var log = require( '../../lib/logger' );
+var log = require( './logger' );
 var serializer = require( 'comment-serializer' );
 var mySerializer = serializer({
     parsers: serializer.parsers()
