+++ lib/js/render.js
@@ -6,10 +6,10 @@ var postcss = require( 'postcss' );
 var prefixer = require( 'postcss-prefix-selector' );
 var Handlebars = require( 'handlebars' );
 
-var hbsHelpers = require( '../../lib/handlebars-helpers' );
-var log = require( '../../lib/logger' );
-var globber = require( '../../lib/globber' );
-var fsp = require( '../../lib/fs-promiser' );
+var hbsHelpers = require( './handlebars-helpers' );
+var log = require( './logger' );
+var globber = require( './globber' );
+var fsp = require( './fs-promiser' );
 
 module.exports = function ( config ) {
