+++ lib/templates/main.hbs
