+++ lib/templates/partials/file-path.hbs
