+++ lib/templates/partials/head.hbs
