+++ lib/templates/partials/masthead.hbs
