+++ lib/templates/partials/section-color.hbs
