+++ lib/templates/partials/section-typography.hbs
