+++ lib/templates/partials/section-variable.hbs
