+++ lib/templates/styles/preview/defaults.scss
