+++ lib/templates/styles/preview/index.scss
