+++ lib/templates/styles/section-color/768.scss
