+++ lib/templates/styles/section-color/defaults.scss
