+++ lib/templates/styles/section-color/index.scss
