+++ lib/templates/styles/section-default/defaults.scss
