+++ lib/templates/styles/section-default/index.scss
