+++ lib/templates/styles/section-typography/defaults.scss
