+++ lib/templates/styles/section-typography/index.scss
