+++ lib/templates/styles/section-variable/defaults.scss
