+++ lib/templates/styles/section-variable/index.scss
