+++ lib/templates/styles/tag-details/defaults.scss
