+++ lib/templates/styles/tag-details/index.scss
